IOUP and Exponential prior handle negative timesteps now
