+++ tests/test_probdiffeq/test_priors/test_wiener_integrated.py
@@ -8,48 +8,15 @@ from probdiffeq.backend import np, testing
     "ssm_factory",
     [probdiffeq.state_space_model_isotropic, probdiffeq.state_space_model_dense],
 )
-@testing.parametrize("sign", [1, -1])
-def test_transitions_are_correct_in_1d_unit_step(ssm_factory, sign) -> None:
-    """Assert that the IWP transition matrix and noise covariance match the analytical formulas in 1d."""
-    tcoeffs = [2.0, 3.0, 4.0, 5.0]
-    ssm = ssm_factory()
-    iwp = ssm.prior_wiener_integrated(tcoeffs)
-
-    cond = iwp.transition(dt=sign * 1.0, output_scale=1.0)
-    A_expected = np.asarray(
-        [
-            [1.0, 3.0, 3.0, 1.0],
-            [0.0, 1.0, 2.0, 1.0],
-            [0.0, 0.0, 1.0, 1.0],
-            [0.0, 0.0, 0.0, 1.0],
-        ]
-    )
-    Q_expected = np.asarray(
-        [
-            [1.0 / 7.0, 1.0 / 6.0, 1.0 / 5.0, 1.0 / 4.0],
-            [1.0 / 6.0, 1.0 / 5.0, 1.0 / 4.0, 1.0 / 3.0],
-            [1.0 / 5.0, 1.0 / 4.0, 1.0 / 3.0, 1.0 / 2.0],
-            [1.0 / 4.0, 1.0 / 3.0, 1.0 / 2.0, 1.0 / 1.0],
-        ]
-    )
-    _, Q_received = cond.noise.to_multivariate_normal()
-    assert testing.allclose(cond.A, A_expected)
-    assert testing.allclose(Q_received, Q_expected)
-
-
-@testing.parametrize(
-    "ssm_factory",
-    [probdiffeq.state_space_model_isotropic, probdiffeq.state_space_model_dense],
-)
 @testing.parametrize("dt", [1.234, -1.234])
-def test_transitions_are_correct_in_1d_nonunit_step(ssm_factory, dt) -> None:
+def test_transitions_are_correct_in_1d(ssm_factory, dt) -> None:
     """Assert that the IWP transition matrix and noise covariance match the analytical formulas in 1d."""
     tcoeffs = [2.0, 3.0, 4.0, 5.0]
     ssm = ssm_factory()
     iwp = ssm.prior_wiener_integrated(tcoeffs)
     cond = iwp.transition(dt=dt, output_scale=1.0).preconditioner_apply()
 
-    # A(dt)_{ij} = dt^{j-i} / (j-i)!   (signed dt; backward step inverts automatically)
+    # A(dt)_{ij} = dt^{j-i} / (j-i)!   (depends on signed dt)
     A_expected = np.asarray(
         [
             [1.0, dt, dt**2 / 2.0, dt**3 / 6.0],
@@ -61,7 +28,7 @@ def test_transitions_are_correct_in_1d_nonunit_step(ssm_factory, dt) -> None:
 
     assert testing.allclose(cond.A, A_expected)
 
-    # Q(dt)_{ij} = |dt|^{7-i-j} / ((3-i)! (3-j)! (7-i-j))   (magnitude; PSD for both signs)
+    # Q(dt)_{ij} = |dt|^{2*q+1-i-j} / ((q-i)! (q-j)! (2*q+1-i-j)) (depends on |dt|)
     h = abs(dt)
     Q_expected = np.asarray(
         [
@@ -72,44 +39,15 @@ def test_transitions_are_correct_in_1d_nonunit_step(ssm_factory, dt) -> None:
         ]
     )
 
-    _, Q_received = cond.noise.to_multivariate_normal()
+    q_received, Q_received = cond.noise.to_multivariate_normal()
+    assert testing.allclose(q_received, np.zeros((4,)))
     assert testing.allclose(Q_received, Q_expected)
 
 
-# Separate test because conditional shapes differ
-@testing.parametrize("sign", [1, -1])
-def test_transitions_are_correct_in_1d_blockdiag_unit_step(sign) -> None:
-    """Assert the same for the blockdiag model, where the conditional has an extra leading batch dimension."""
-    tcoeffs = [2.0, 3.0, 4.0, 5.0]
-    ssm = probdiffeq.state_space_model_blockdiag()
-    iwp = ssm.prior_wiener_integrated(tcoeffs)
-
-    cond = iwp.transition(dt=sign, output_scale=np.ones((1,)))
-    A_expected = np.asarray(
-        [
-            [1.0, 3.0, 3.0, 1.0],
-            [0.0, 1.0, 2.0, 1.0],
-            [0.0, 0.0, 1.0, 1.0],
-            [0.0, 0.0, 0.0, 1.0],
-        ]
-    )
-    Q_expected = np.asarray(
-        [
-            [1.0 / 7.0, 1.0 / 6.0, 1.0 / 5.0, 1.0 / 4.0],
-            [1.0 / 6.0, 1.0 / 5.0, 1.0 / 4.0, 1.0 / 3.0],
-            [1.0 / 5.0, 1.0 / 4.0, 1.0 / 3.0, 1.0 / 2.0],
-            [1.0 / 4.0, 1.0 / 3.0, 1.0 / 2.0, 1.0 / 1.0],
-        ]
-    )
-    assert testing.allclose(cond.A, A_expected[None, ...])
-    cov = np.einsum("ijk,ilk->ijl", cond.noise.cholesky_flat, cond.noise.cholesky_flat)
-    assert testing.allclose(cov, Q_expected[None, ...])
-
-
-# Separate test because conditional shapes differ
+# Separate test for blockdiag because output-scale shapes differ
 @testing.parametrize("dt", [1.234, -1.234])
-def test_transitions_are_correct_in_1d_blockdiag_nonunit_step(dt) -> None:
-    """Assert the same for the blockdiag model, where the conditional has an extra leading batch dimension."""
+def test_transitions_are_correct_in_1d_blockdiag(dt) -> None:
+    """Assert that the IWP transition matrix and noise covariance match the analytical formulas in 1d."""
     tcoeffs = [2.0, 3.0, 4.0, 5.0]
     ssm = probdiffeq.state_space_model_blockdiag()
     iwp = ssm.prior_wiener_integrated(tcoeffs)
@@ -138,7 +76,6 @@ def test_transitions_are_correct_in_1d_blockdiag_nonunit_step(dt) -> None:
 
     assert testing.allclose(cond.A, A_expected[None, ...])
 
-    _, cov = cond.noise.to_multivariate_normal()
-    print(cov)
-    print(Q_expected)
-    assert testing.allclose(cov, Q_expected)
+    q_received, Q_received = cond.noise.to_multivariate_normal()
+    assert testing.allclose(q_received, np.zeros((4,)))
+    assert testing.allclose(Q_received, Q_expected)
